+++ cmake/Rust.cmake
@@ -1,3 +1,4 @@
+include(FeatureSummary)
 include(FindRust)
 find_package(Rust REQUIRED)
 
@@ -20,10 +21,17 @@ endif()
 
 set(rust_profile $<IF:$<CONFIG:Debug>,debug,$<IF:$<CONFIG:RelWithDebInfo>,release-with-debug,release>>)
 
-option(WITH_GETTEXT "Build with gettext localization support. Requires `msgfmt` to work." ON)
+if (NOT DEFINED WITH_MESSAGE_LOCALIZATION) # Don't check for legacy options if the new one is defined, to help bisecting.
+    if(DEFINED WITH_GETTEXT)
+        message(FATAL_ERROR "the WITH_GETTEXT option is no longer supported, use -DWITH_MESSAGE_LOCALIZATION=ON|OFF")
+    endif()
+endif()
+option(WITH_MESSAGE_LOCALIZATION "Build with localization support. Requires `msgfmt` to work." ON)
 # Enable gettext feature unless explicitly disabled.
-if(NOT DEFINED WITH_GETTEXT OR "${WITH_GETTEXT}")
+if(NOT DEFINED WITH_MESSAGE_LOCALIZATION OR "${WITH_MESSAGE_LOCALIZATION}")
     list(APPEND FISH_CARGO_FEATURES_LIST "localize-messages")
 endif()
+
+add_feature_info(Translation WITH_MESSAGE_LOCALIZATION "message localization (requires gettext)")
 
 list(JOIN FISH_CARGO_FEATURES_LIST , FISH_CARGO_FEATURES)